weather/requirements.txt
requests>=2.31.0
schedule>=1.2.1
pytz>=2024.1
urllib3>=2.2.1
python-dotenv>=1.0.0